// src/testing-system/table/Table.types.ts
// export interface HardwareIOInterface {
//     Comments : string;
//     ExerciseID: number;
//     ExerciseName : string;
//     FlightNumber : number;
//     Name : string;
//     Title: string;
//     Value : string;
//     UpdateCounter: number;
//     UpdateTime : string;
//     Group: string
//     id : number;
//     HeightASL: string;
//     HeightMSL: string;
//     Fuel: string;
//     RateOfClimb: string;
//     Velocity: string;
//     CrashReason: string;

//     _id : string;
// }

export interface AutoTestPlanListItem {
    test_id : number;
    test_name: string;
    main_version: string;
    test_prerequisite: string;
    test_environment: string;
    test_threshold: string;
    test_runtime: string;
    test_site: string;
    test_summary: string;
}




// export interface HardwareIOListInterface {
//     data : HardwareIOInterface[];
//     success : boolean;
// }